refactor(access): Update access imports to neew path
ref: #640 #641
