+++ app/access/admin.py
@@ -2,7 +2,10 @@ from django.contrib import admin
 from django.contrib.auth.models import Group, User
 from django.contrib.auth.admin import UserAdmin
 
-from .models import *
+
+from access.models.organization import Organization
+from access.models.team import Team
+from access.models.team_user import TeamUsers
 
 admin.site.unregister(Group)
