+++ app/access/forms/organization.py
@@ -3,7 +3,7 @@ from django.db.models import Q
 
 from app import settings
 
-from access.models import Organization
+from access.models.organization import Organization
 
 from core.forms.common import CommonModelForm
