+++ app/access/serializers/team_user.py
@@ -2,7 +2,7 @@ from rest_framework.reverse import reverse
 
 from rest_framework import serializers
 
-from access.models import TeamUsers
+from access.models.team_user import TeamUsers
 
 from api.serializers import common
