+++ app/access/tests/functional/team_user/test_team_user_permission_viewset.py
@@ -10,7 +10,9 @@ from django.test import Client, TestCase
 
 from rest_framework.reverse import reverse
 
-from access.models import Organization, Team, TeamUsers
+from access.models.organization import Organization
+from access.models.team import Team
+from access.models.team_user import TeamUsers
 
 from api.tests.abstract.test_metadata_functional import MetadataAttributesFunctional
 from api.tests.abstract.api_permissions_viewset import APIPermissions
