+++ app/access/tests/functional/team_user/test_team_user_serializer.py
@@ -6,7 +6,8 @@ from django.test import TestCase
 
 from rest_framework.exceptions import ValidationError
 
-from access.models import Organization, Team
+from access.models.organization import Organization
+from access.models.team import Team
 
 from access.serializers.team_user import (
     TeamUsers,
