+++ app/access/tests/unit/organization/test_organizaiton_api_v2.py
@@ -9,7 +9,10 @@ from django.test import Client, TestCase
 
 from rest_framework.relations import Hyperlink
 
-from access.models import Organization, Team, TeamUsers
+from access.models.organization import Organization
+from access.models.team import Team
+from access.models.team_user import TeamUsers
+
 
 from api.tests.abstract.api_fields import APICommonFields
