+++ app/access/tests/unit/organization/test_organizaiton_permission_api.py
@@ -7,7 +7,9 @@ from django.contrib.contenttypes.models import ContentType
 from django.shortcuts import reverse
 from django.test import Client, TestCase
 
-from access.models import Organization, Team, TeamUsers
+from access.models.organization import Organization
+from access.models.team import Team
+from access.models.team_user import TeamUsers
 
 from api.tests.abstract.api_permissions import APIPermissionChange, APIPermissionView
