+++ app/access/tests/unit/organization/test_organization_viewset_unit.py
@@ -5,7 +5,8 @@ from django.test import Client, TestCase
 
 from rest_framework.reverse import reverse
 
-from access.models import Organization
+
+from access.models.organization import Organization
 from access.viewsets.organization import ViewSet
 
 from api.tests.abstract.viewsets import ViewSetModel
