+++ app/access/tests/unit/organization_notes/test_unit_organization_notes_viewset.py
@@ -5,7 +5,7 @@ from django.test import Client, TestCase
 
 from rest_framework.reverse import reverse
 
-from access.models import Organization
+from access.models.organization import Organization
 from access.viewsets.team_notes import ViewSet
 
 from api.tests.abstract.viewsets import ViewSetModel
