+++ app/access/tests/unit/team/test_team_api.py
@@ -11,7 +11,9 @@ from django.test import Client, TestCase
 
 from rest_framework.relations import Hyperlink
 
-from access.models import Organization, Team, TeamUsers
+from access.models.organization import Organization
+from access.models.team import Team
+from access.models.team_user import TeamUsers
 
 # from api.tests.abstract.api_permissions import APIPermissions
